feat : 튜티, 카테고리 update api 리턴 타입 변경
